Add -a (accurate capture mode) flag for capturing frames in accurate mode using ffmpeg.
This is especially useful for HEVC videos where captured frames look terrible when using fast capture mode.

The downside is that accurate mode is much slower. This is due to the fact that we have to decode multiple frames before getting the actual frame that we want in order to get good results.

The number of seconds that are decoded before each captured frame can be specified with the -A flag.