Fix a thin line of background showing on the right side of contact sheets
If the contact sheet width (minus horizontal spacing) is not evenly divisible
by the number of image columns, a thin line of background is left visible on
the right side after composition.

Fix this by interpreting the value of --width as a target we try to match as
closely as possible, instead of as an unchangable constant.
In the current implementation this is an upper limit.