add `--fast` option for generating contact sheets as fast as possible regardless of output image quality.
When `--fast` is enabled, blurriness and average color is not computed for each frame and frame capture is multithreaded.

The current implementation appears to mess up the terminal after vcsi finishes executing. Typing `reset` or `stty sane` fixes it.

This is an option for now because it is not stable enough, but may become the default at a later point.