Improve handling of non-zero ffmpeg responses
Currently 'make_capture()' just drops the output from the ffmpeg CLI.

By switching to 'subprocess.run()' it is possible to handle non-zero
exit codes from the ffmpeg CLI. This prints out the full ffmpeg command,
exit code and the response from ffmpeg which should be enough for a
tech-savvy user to debug.

Given this will nearly always be an issue with the ffmpeg installation
or the file there isn't anything more vcsi can do in this situation.

This is implemented as an exception so that when run with
'--ignore-errors' the file can be skipped.

Fixes: https://github.com/amietn/vcsi/issues/104
Fixes: https://github.com/amietn/vcsi/issues/103
Fixes: https://github.com/amietn/vcsi/issues/86
Fixes: https://github.com/amietn/vcsi/issues/82
Fixes: https://github.com/amietn/vcsi/issues/75